Sandbox sync updates
